test PHP CGI get and post
