+++ web_pages/cgi_bin/hello_get_PHP.html
@@ -1,0 +1,5 @@
+<form action = "./hello.php" method = "get">
+    First Name: <input type = "text" name = "first_name">  <br />
+    Last Name: <input type = "text" name = "last_name" />
+    <input type = "submit" value = "Submit" />
+</form>
